refactor: clean code follow standard format
