+++ auth-service/src/main/java/avg/vnlaw/authservice/config/security/SecurityConfig.java
@@ -5,7 +5,7 @@ import avg.vnlaw.authservice.o2auth.handler.OAuth2AuthenticationFailureHandler;
 import avg.vnlaw.authservice.o2auth.handler.OAuth2AuthenticationSuccessHandler;
 import avg.vnlaw.authservice.o2auth.services.OAuth2Service;
 import avg.vnlaw.authservice.services.LogoutService;
-import avg.vnlaw.authservice.services.impl.UserService;
+import avg.vnlaw.authservice.services.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
