+++ auth-service/src/main/java/avg/vnlaw/authservice/services/ReCaptchaService.java
@@ -1,4 +1,4 @@
-package avg.vnlaw.authservice.services.impl;
+package avg.vnlaw.authservice.services;
 
 import avg.vnlaw.authservice.responses.ReCaptchaResponse;
 import io.github.cdimascio.dotenv.Dotenv;
