+++ auth-service/src/main/java/avg/vnlaw/authservice/services/TokenService.java
@@ -1,17 +1,17 @@
-package avg.vnlaw.authservice.services.impl;
+package avg.vnlaw.authservice.services;
 
 import avg.vnlaw.authservice.entities.Token;
 import avg.vnlaw.authservice.entities.User;
 import avg.vnlaw.authservice.repositories.TokenRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
+@RequiredArgsConstructor
 public class TokenService {
-    @Autowired
-    private TokenRepository tokenRepository;
 
-    
+    private final TokenRepository tokenRepository;
+
     public void saveUserToken(User user, String jwtToken) {
         var token = Token.builder()
                 .user(user)
@@ -25,8 +25,8 @@ public class TokenService {
     
     public void revokedAllUserTokens(User user) {
         var validUserTokens = tokenRepository.findAllValidTokenByUser(user.getId());
-        if(validUserTokens.isEmpty())
-            return;
+        if(validUserTokens.isEmpty()) return;
+
         validUserTokens.forEach(token -> {
             token.setExpired(true);
             token.setRevoked(true);
